==> 03 Scripts/1. Data Wrangling and Subsetting.ipynb <==
# 1. Importing libraries

import pandas as pd
import numpy as np
import os

# 2. Importing data

#create path
path = r'C:\Users\Mathilde\Documents\DATA ANALYSIS CAREERFOUNDRY\May 2024 - Instacart Basket Analysis'
#orders data
df_ords = pd.read_csv(os.path.join(path, '02 Data', 'Original data', 'orders.csv'), index_col = False)
#products data
df_prods = pd.read_csv(os.path.join(path, '02 Data', 'Original data', 'products.csv'), index_col = False)
#departments data
df_dep = pd.read_csv(os.path.join(path, '02 Data', 'Original data', 'departments.csv'), index_col = False)

# 3. Wrangling data

## 3.1. Orders data frame

#dropping eval_set column from orders data frame
df_ords=df_ords.drop(columns = ['eval_set'])

#changing data type for order and user ID in orders data frame
df_ords['order_id'] = df_ords['order_id'].astype('str')
df_ords['user_id'] = df_ords['user_id'].astype('str')

## 3.2. Departments data frame

#transposing df_dep
df_dep_t=df_dep.T

### Change header of departments df

#take 1st row of df_dep_t for the header
new_header = df_dep_t.iloc[0]

#copying df_dep_t from row 1 onward
df_dep_t_new = df_dep_t[1:]

#set the header row as the df header
df_dep_t_new.columns = new_header

### Checking result of the transformation

df_dep_t_new

#### Observation:

The data frame is no longer wide and the header has been fixed successfully.

# 4. Creating Data Dictionary

data_dict = df_dep_t_new.to_dict('index')

# 5. Task

# Step 2: it doesn't make sense to make stats on the order number as it is basically an ID (for each user, the order number is unique) so I am changing it to string.
df_ords['order_number'] = df_ords['order_number'].astype('str')

#checking if the type changed - doesn't appear in the stats so aok
df_ords.describe()

#Step 3: rename column day of the week in orders data frame (there is no other unintuitive name in the set so I used the same as in the exercise)
df_ords.rename(columns = {'order_dow' : 'orders_day_of_the_week'}, inplace = False)

#Step 4: finding busiest hour
df_ords['order_hour_of_day'].value_counts(dropna = True)

#### Observation:

The busiest time of the day to place orders is 10am, closely followed by 11am.

#Step 5: what "4" means in the data dictionary
print(data_dict.get('4'))

#### Observation:

The department_id "4" corresponds to the category "produce".

#Step 6: Creating subset for breakfast items. First, identify breakfast dept id:
data_dict

-> department_id for "breakfast" is 14.

#creating the subset
df_prods_breakfast = df_prods.loc[df_prods['department_id'] == 14]

#checking result
df_prods_breakfast.head()

Step 7: department_id for "alcohol", "deli", "bevrages" and "meat/seafood" are respectively 5, 20, 7 and 12.

#Step 7: creating "dinner party" subset.
df_prods_dinner_party = df_prods.loc[df_prods['department_id'].isin([5,7,12,20])]

#### Observation:

The sales team can use the data frame df_prods_dinner_party for their analysis.

#Step 8: find the number of rows
df_prods_dinner_party.shape

#### Observation:

The new dinner party data frame has 7650 rows.

#Step 9: finding info about customer with user_id 1
df_user1=df_ords.loc[df_ords['user_id'] == '1']
df_user1

#Step 10: stats about the user's behaviour
df_user1.describe()

#### Observation:

This user only places orders from Sunday to Wednesday, never from Thrusday to Saturday. More than half of the time, he places orders at 8am or earlier. He places orders on average every 19 days. He never spend more than 30 days without placing orders , has already placed 2 orders the same day.

# 6. Export Data

#export orders data frame
df_ords.to_csv(os.path.join(path, '02 Data','Prepared Data', 'orders_wrangled.csv'))

#export departments data frame
df_dep_t_new.to_csv(os.path.join(path, '02 Data','Prepared Data', 'departments_wrangled.csv'))
