==> 03 Scripts/2. Data Consistency Checks.ipynb <==
# 1. Importing libraries

import pandas as pd
import numpy as np
import os

# 2. Importing data

#create path
path = r'C:\Users\Mathilde\Documents\DATA ANALYSIS CAREERFOUNDRY\May 2024 - Instacart Basket Analysis'
#orders data
df_ords = pd.read_csv(os.path.join(path, '02 Data', 'Prepared data', 'orders_wrangled.csv'), index_col = False)
#products data
df_prods = pd.read_csv(os.path.join(path, '02 Data', 'Original data', 'products.csv'), index_col = False)
#departments data
df_dep = pd.read_csv(os.path.join(path, '02 Data', 'Prepared data', 'departments_wrangled.csv'), index_col = False)

import warnings
warnings.filterwarnings('ignore')

# 3. Data consistency check on Products dataframe

#Checking for missing values
df_prods.isnull().sum()

#View the missing values
df_nan=df_prods[df_prods['product_name'].isnull()==True]
df_nan

#Creata a data frame without the missing values
df_prods_clean=df_prods[df_prods['product_name'].isnull()==False]
#checking if the missing values have been removed by viewing size of new df
df_prods_clean.shape

#checking size of original df
df_prods.shape

#viewing duplicates
df_dups = df_prods_clean[df_prods_clean.duplicated()]
df_dups

df_prods_clean.shape

df_prods_clean_no_dups = df_prods_clean.drop_duplicates()

df_prods_clean_no_dups.shape

### Checking for inconsistencies

df_prods.describe()

### Observation:

Maximum value for price is 99999, which is very much higher than the mean (and therefore, the std is also very high). This is not a price expected for a supermarket product. Let's investigate further. Let's consider no product in a supermarket would cost over 100 dollars.

#identify products with abnormally high price
df_prods_clean_no_dups.loc[df_prods['prices'] > 100]

We found a pack of cottage cheese for 14900 and a pack of milk for 99999 dollars, which are very obviouslu the result of typos or data corruption. Let's assume that a comma has been ommited and correct it:

#replacing outliers 
df_prods_clean_no_dups = df_prods_clean_no_dups.replace ({"prices":{99999.0: 9.99, 14900.0:1.49 }})

df_prods_clean_no_dups.describe()

# 4. Changing data type

#changing data type for product_id, aisle_id and department_id from int to string
df_prods_clean_no_dups['product_id'] = df_prods_clean_no_dups['product_id'].astype('str')
df_prods_clean_no_dups['aisle_id'] = df_prods_clean_no_dups['aisle_id'].astype('str')
df_prods_clean_no_dups['department_id']= df_prods_clean_no_dups['department_id'].astype('str')

# 5. Export cleaned product data

#export cleaned prods data frame
df_prods_clean_no_dups.to_csv(os.path.join(path, '02 Data','Prepared Data', 'products_checked.csv'))

# 6. Task - Data consistency check on Orders data frame

#Dropping the useless column that showed up
df_ords=df_ords.drop(columns=['Unnamed: 0'])

## 6.2 Checking for inconsistencies in Orders DF

df_ords.describe()

### Observation:

Day of the week ranges from 0 to 6 - this is normal, 7 days in a week.
Hour of the day ranges from 0 to 23 - also expected
Days since prior order ranges from 0 to 30 - also seems coherent.

## 6.3 Checking for mixed-type data in Orders DF

for col in df_ords.columns.tolist():
  weird = (df_ords[[col]].applymap(type) != df_ords[[col]].iloc[0].apply(type)).any(axis = 1)
  if len (df_ords[weird]) > 0:
    print (col)

### Observation:

No column returned means that there is no mix-typed data.

## 6.5 Checking for missing values in Orders DF

df_ords.isnull().sum()

### Observation:

It is expected to have missing values in the column "days since prior order": every customer has a first order, for which there is no prior order. The number of missing values should actually match the number of customers (as each customer has one first order)

df_ords.tail(5)

The last user_id is 206209, meaning there are 206209 customers: which is exactly the number of missing values. This is coherent. No action taken.

I found another way of checking this, as the column user_id might not be sorted in descending order:

df_ords['user_id'].nunique()

## 6.7 Checking for duplicates in Orders DF

#view duplicates
df_ords[df_ords.duplicated()]

### Observation:

The resulting view is empty: no duplicates. No action taken

###### No change made to the orders data frame, so I am not exporting it.

df_ords.shape
