==> 03 Scripts/3. Combining and Exporting Data Part1.ipynb <==
# 1. Importing libraries

import pandas as pd
import numpy as np
import os

# 2. Importing data

#create path
path = r'C:\Users\Mathilde\Documents\DATA ANALYSIS CAREERFOUNDRY\May 2024 - Instacart Basket Analysis'
#orders data
vars_list = ['order_id', 'user_id', 'order_number', 'order_dow', 'order_hour_of_day', 'days_since_prior_order']
df_ords = pd.read_csv(os.path.join(path, '02 Data', 'Prepared data', 'orders_wrangled.csv'), usecols = vars_list)
#orders prior
df_ords_prior = pd.read_csv(os.path.join(path, '02 Data', 'Original data', 'orders_products_prior.csv'), index_col = False)

#exploring the new dataframe
df_ords_prior.head()

df_ords_prior.shape

df_ords_prior.describe()

# 3. Merging Orders dataframe with orders product prior dataframe

df_merged_large = df_ords.merge(df_ords_prior, on = 'order_id', indicator = True)

#checking result
df_merged_large.head(20)

df_merged_large['_merge'].value_counts()

#checking shape
df_merged_large.shape

# 4. Export new data frame

#export as pickle
df_merged_large.to_pickle(os.path.join(path, '02 Data','Prepared Data', 'orders_products_combined.pkl'))
