==> 03 Scripts/4. Combining and Exporting Data Part2.ipynb <==
# 1. Importing libraries

import pandas as pd
import numpy as np
import os

# 2. Importing data

#create path
path = r'C:\Users\Mathilde\Documents\DATA ANALYSIS CAREERFOUNDRY\May 2024 - Instacart Basket Analysis'
#orders product combined data
df_ords_prods = pd.read_pickle(os.path.join(path, '02 Data', 'Prepared data', 'orders_products_combined.pkl'))

#products data
df_prods = pd.read_csv(os.path.join(path, '02 Data', 'Prepared data', 'products_checked.csv'), index_col = False)

#checking the new import
df_ords_prods.shape

# 3. Combining orders product prior dataframe w/ products data frame

#Visualize the dataframes to help
df_ords_prods.head()

#Need to drop the column "_merge" because it's giving me an error when trying to merge with indicator in the next step
df_ords_prods=df_ords_prods.drop(columns = ['_merge'])

#checking if column was removed
df_ords_prods.head()

df_prods.head()

### Reflexion on the combination:

The two dataframes are not the same shape, they can't be concatenated.

The two dataframes have one column in common: product_id. It will be our key for merging

We're using an inner join in order to only get products that appear in both data sets.

df_merged = df_ords_prods.merge(df_prods, on='product_id', how='inner', indicator=True)

#merge flag
df_merged['_merge'].value_counts() 

#dropping the "_merge" column because I won't need it anymore - sparing a bit of memory
df_merged=df_merged.drop(columns = ['_merge'])

### Observation:

The only value in "merge" column is "both" - which is to be expexted as we used an inner join. We got rid of the 2029 unwanted lines that we got with the left join, from the 2 products that were not referenced in the prods data set.

# 4. Export new data frame

#export as pickle, as new data set is very large, csv would be to slow to export and read
df_merged.to_pickle(os.path.join(path, '02 Data','Prepared Data', 'ords_prods_merge.pkl'))
