==> 03 Scripts/5. Deriving New Variables.ipynb <==
#importing libraries
import pandas as pd
import numpy as np
import os

#importing data
#create path
path = r'C:\Users\Mathilde\Documents\DATA ANALYSIS CAREERFOUNDRY\May 2024 - Instacart Basket Analysis'
#orders product combined data
df_ords_prods = pd.read_pickle(os.path.join(path, '02 Data', 'Prepared data', 'ords_prods_merge.pkl'))

# Create price labels

#defining function
def price_label(row):

  if row['prices'] <= 5:
    return 'Low-range product'
  elif (row['prices'] > 5) and (row['prices'] <= 15):
    return 'Mid-range product'
  elif row['prices'] > 15:
    return 'High range'
  else: return 'Not enough data'

df_ords_prods.loc[df_ords_prods['prices'] > 15, 'price_range_loc'] = 'High-range product'

df_ords_prods.loc[(df_ords_prods['prices'] <= 15) & (df_ords_prods['prices'] > 5), 'price_range_loc'] = 'Mid-range product' 

df_ords_prods.loc[df_ords_prods['prices'] <= 5, 'price_range_loc'] = 'Low-range product'

df_ords_prods['price_range_loc'].value_counts()

# Create labels for busy days

#identifying busy days and least busy days
df_ords_prods['order_dow'].value_counts(dropna=False)

### Observation

The most busy day is 0 (Sunday). Second most busy is 1 (Monday). Least busy is 4 (Wednesday). Second least busy is 3 (Tuesday).

#labelling busy days
result = []

for value in df_ords_prods["order_dow"]:
  if value == 0:
    result.append("Busiest days")
  elif value == 4:
    result.append("Least busy")
  else:
    result.append("Regularly busy")

#adding the resulting list to the data frame
df_ords_prods['busiest_day'] = result

#frequency of values
df_ords_prods['busiest_day'].value_counts(dropna=False)

The count of values seem to make sense - the 2 busiest days gat almost as products ordered than the 3 regularly busy days. And the least busy days get much fewer products ordered, also expected, as they're less busy.

# Create labels for busy hours

#identifying busy hours and less busy hours
df_ords_prods['order_hour_of_day'].value_counts(dropna=False)

### Observation

We could identify more or less large periods of time (for exemple, top 2 busiest hours, or top 8 etc.)
I was thibking of taking top 4 or top 5, but it wouldn't work, as it would create 2 distinct periods (10-11 and then 13-15, or 14-15.) So I'll go for top 6 hours and get the uninterrupted period 10-15 (10am-3pm). It also works for the bottom 6, getting the low period of 0-5 (12am - 5am)

#labelling busy periods
result2 = []

for value in df_ords_prods["order_hour_of_day"]:
  if value >= 10 and value <= 15:
    result2.append("Most orders")
  elif value >= 0 and value <= 5:
    result2.append("Fewest orders")
  else:
    result2.append("Average orders")

#adding the resulting list to the data frame
df_ords_prods['busiest_period_of_day'] = result2

#frequency of values
df_ords_prods['busiest_period_of_day'].value_counts(dropna=False)

# Exporting new data frame

df_ords_prods.to_pickle(os.path.join(path, '02 Data','Prepared Data', 'ords_prods_merge_derived.pkl'))
