==> 03 Scripts/6. Grouping Data and Aggregating Variables.ipynb <==
#importing libraries
import pandas as pd
import numpy as np
import os

#importing data
#create path
path = r'C:\Users\Mathilde\Documents\DATA ANALYSIS CAREERFOUNDRY\May 2024 - Instacart Basket Analysis'
#orders product combined data
ords_prods_merge = pd.read_pickle(os.path.join(path, '02 Data', 'Prepared data', 'ords_prods_merge_derived.pkl'))

#ignore warnings
import warnings
warnings.filterwarnings('ignore')

# Aggregating and calculating means

#creating subset
df = ords_prods_merge[:1000000]

#calculate the mean of the subset by aggregating
df.groupby('department_id').agg({'order_number': ['mean']})

#calculate the mean of the entiere dataframe by aggregating
ords_prods_merge.groupby('department_id').agg({'order_number': ['mean']})

### Observation:

The means for each department are similar, but not exactly the same, which was to be expected.

# Creating loyalty flag

#create a new column with the max orders for each user
ords_prods_merge['max_order'] = ords_prods_merge.groupby(['user_id'])['order_number'].transform(np.max)

#checking result
ords_prods_merge.head(15)

#creation of loyalty flag
ords_prods_merge.loc[ords_prods_merge['max_order'] > 40, 'loyalty_flag'] = 'Loyal customer'
ords_prods_merge.loc[(ords_prods_merge['max_order'] <= 40) & (ords_prods_merge['max_order'] > 10), 'loyalty_flag'] = 'Regular customer'
ords_prods_merge.loc[ords_prods_merge['max_order'] <= 10, 'loyalty_flag'] = 'New customer'

#checking result
ords_prods_merge.head(1)

#Frequency of loyalty flag
ords_prods_merge['loyalty_flag'].value_counts(dropna = False)

### Observation:

It is my understanding that the frequency check doesn't indicate the number of new/regular/loyal customers: it indicates, for each individual product in each individual order, if the product has been ordered by a new/regular/loyal customer. Which is why naturally the frequency of "new customer" is much lower. I think it would be interested to transform the data frame, grouping by user_id and only displaying the loyalty flag once for each user_id: then we could know how many customers are new/reg/loyal.

# Spending habits of customers

Note: when trying to execute the code below, I had a memory error and my computer crashed. So I'm dropping a useless column to help with this issue.

ords_prods_merge = ords_prods_merge.drop(columns=['Unnamed: 0'])

#calculating mean, min and max of prices of products ordered by each loyalty category
ords_prods_merge.groupby('loyalty_flag').agg({'prices': ['mean', 'min', 'max']})

### Observation:

There seem to be no significant different in the consumption habit of the different types of customers. Min and max are equal, and the means are very, very close.

# Creating spending flag

#create a new column with the average price
ords_prods_merge['avg_price'] = ords_prods_merge.groupby(['user_id'])['prices'].transform(np.mean)

#check result
ords_prods_merge.head(5)

#creation of spending flag
ords_prods_merge.loc[ords_prods_merge['avg_price'] < 10, 'spending_flag'] = 'Low spender'
ords_prods_merge.loc[(ords_prods_merge['avg_price'] >= 10), 'spending_flag'] = 'High spender'

#check result
ords_prods_merge.head(5)

#Frequency of spending flag
ords_prods_merge['spending_flag'].value_counts(dropna = False)

### Observation:

There are a lot more low spenders than high spenders (or, most of the products ever bought have been bought by low spenders rather than high spenders)

# Creating order frequency flag

#create a new column with the median number of days between orders
ords_prods_merge['median_dspo'] = ords_prods_merge.groupby(['user_id'])['days_since_prior_order'].transform(np.median)

#check result
ords_prods_merge.head(5)

#creation of order frequency flag
ords_prods_merge.loc[ords_prods_merge['median_dspo'] > 20, 'order_frequency_flag'] = 'Non-frequent customer'
ords_prods_merge.loc[(ords_prods_merge['median_dspo'] <= 20) & (ords_prods_merge['median_dspo'] > 10), 'order_frequency_flag'] = 'Regular customer'
ords_prods_merge.loc[ords_prods_merge['median_dspo'] <= 10, 'order_frequency_flag'] = 'Frequent customer'

#check result
ords_prods_merge.head(5)

#Frequency of order frequency flag
ords_prods_merge['order_frequency_flag'].value_counts(dropna = False)

"NaN" value accounts for customers that placed only one order.

# Exporting data

ords_prods_merge.to_pickle(os.path.join(path, '02 Data','Prepared Data', 'ords_prods_merge_derived_w_flags.pkl'))
