==> 03 Scripts/9. Exclusing low activity customers.ipynb <==
# Import libraries and data

#importing libraries
import pandas as pd
import numpy as np
import os

#importing data
#create path
path = r'C:\Users\Mathilde\Documents\DATA ANALYSIS CAREERFOUNDRY\May 2024 - Instacart Basket Analysis'
#final merged data set
df = pd.read_pickle(os.path.join(path, '02 Data', 'Prepared data', 'ords_prods_customers.pkl'))

# Security implications:


We deal with personal cultomer information (names that can be linked to shopping habits) and we have to comply to data privacy regulations. We dropped the columns containing the names of the customers: we "scrubbed" this information out to avoid any issue regarding personally identifiable information (PII)

# Excluding low activity customers

# Creation of a new data frame dropping all the rows corresponding to customers with fewer than 5 orders (we're not interested in them because they don't generate much revenue)
df2 = df.drop(df[df.max_order < 5].index)

# Checking result 
df2['max_order'].describe()

# Exportation of the new df
df2.to_pickle(os.path.join(path, '02 Data', 'Prepared Data', 'ords_prods_merge_high_activ_cx.pkl'))
